flask-server/requirements.txt
flask
opencv-python
matplotlib
ultralytics
numpy
torch
torchvision
Pillow
flask_cors
gunicorn
gevent
